feat: ⑦ 파이프라인 GCP 전환 — BigQuery 단일화 + Dataproc Serverless 배치
- Silver: GCS parquet + BigQuery External Table(hive tx_date 파티션)로 조회
- Gold: dbt-bigquery(fraud_gold) 단일화, DuckDB 완전 제거(임시 대체재 정리)
- DAG spark_silver: DataprocCreateBatchOperator로 전환(+코드 GCS 업로드 태스크)
- reconcile/push_metrics: google-cloud-bigquery 클라이언트로 재작성
- airflow 이미지에 apache-airflow-providers-google 추가
- 검증: Dataproc batch SUCCEEDED, dbt 4모델/reconcile 890==890 정합성 통과

+++ dbt/models/gold/account_risk.sql
@@ -9,7 +9,7 @@ SELECT
     SUM(CASE WHEN is_suspicious THEN 1 ELSE 0 END)     AS suspicious_tx_count,
     SUM(CASE WHEN isFraud = 1 THEN amount ELSE 0 END)  AS fraud_amount,
     ROUND(
-        SUM(CASE WHEN isFraud = 1 THEN 1 ELSE 0 END)::DOUBLE / COUNT(*) * 100, 2
+        SUM(CASE WHEN isFraud = 1 THEN 1 ELSE 0 END) / COUNT(*) * 100, 2
     )                                                  AS fraud_rate_pct,
     MAX(tx_timestamp)                                  AS last_tx_at
 FROM {{ ref('stg_silver_transactions') }}
